python/requirements.txt
flask 
sqlalchemy
mysql-connector-python
pandas
scikit-learn
python-dotenv
flask-cors
joblib
matplotlib
numpy
geopy
googlemaps
openrouteservice
fastapi
uvicorn